requirements.txt
streamlit==1.32.0
PyMuPDF
ebooklib
nltk
tqdm
openai
sentence-transformers
chromadb
faiss-cpu
streamlit-chat
numpy
transformers
torch 